+++ site/client/css/homepage.scss
@@ -1,0 +1,5 @@
+
+
+.thumbnail img {
+    max-width: 5rem;
+}
