+++ site/client/css/variables.scss
@@ -1,0 +1,2 @@
+$sans-serif-font-stack: "Source Sans Pro", "Helvetica Neue", Arial, sans-serif;
+$serif-font-stack: "Merriweather", "Georgia", serif;
